Use the /proc filesystem on cygwin
Cygwin provides a working /proc filesystem that is pretty similar to
Linux distributions. P9Y::ProcessTable::Win32 is updated to recognize
when running on Cygwin and provide the following behavior:

- Use the /proc filesystem to provide Cygwin process info
- Use Win32::Process::Info to provide Windows process info

This provides the user with both Cygwin and Windows processes
when on cygwin and only Windows processes when on Win32.